requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
psycopg2-binary==2.9.7
bcrypt==4.0.1
pandas==2.0.3
xlsxwriter==3.1.3
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0